+++ src/main/java/org/recap/service/submitcollection/SubmitCollectionValidationService.java
@@ -10,6 +10,7 @@ import org.recap.model.report.SubmitCollectionReportInfo;
 import org.recap.repository.jpa.InstitutionDetailsRepository;
 import org.recap.repository.jpa.ItemDetailsRepository;
 import org.recap.service.common.SetupDataService;
+import org.recap.util.CommonUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -44,6 +45,9 @@ public class SubmitCollectionValidationService {
 
     @Autowired
     private SubmitCollectionHelperService submitCollectionHelperService;
+
+    @Autowired
+    private CommonUtil commonUtil;
 
     @Value("${nonholdingid.institution}")
     private String nonHoldingIdInstitution;
@@ -114,14 +118,7 @@ public class SubmitCollectionValidationService {
                         isValid = false;
                         isValidToProcess &= isValid;
                     } else {
-                        SubmitCollectionReportInfo submitCollectionReportInfo = new SubmitCollectionReportInfo();
-                        submitCollectionReportInfo.setItemBarcode(incomingItemEntity.getBarcode());
-                        submitCollectionReportInfo.setCustomerCode(incomingItemEntity.getCustomerCode());
-                        submitCollectionReportInfo.setOwningInstitution(owningInstitution);
-                        String existingOwningInstitutionHoldingsId = getExistingItemEntityOwningInstItemId(fetchedBibliographicEntity,incomingItemEntity);
-                        submitCollectionReportInfo.setMessage(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD+" - Owning institution holdings id mismatch - incoming owning institution holdings id " +incomingHoldingItemMapEntry.getKey()+ ", existing owning institution item id "+incomingItemEntity.getOwningInstitutionItemId()
-                                +", existing owning institution holdings id "+existingOwningInstitutionHoldingsId+", existing owning institution bib id "+fetchedBibliographicEntity.getOwningInstitutionBibId());
-                        failureSubmitCollectionReportInfoList.add(submitCollectionReportInfo);
+                        commonUtil.buildSubmitCollectionReportInfoAndAddFailures(fetchedBibliographicEntity, failureSubmitCollectionReportInfoList, owningInstitution, incomingHoldingItemMapEntry, incomingItemEntity);
                         isValid = false;
                         isValidToProcess &= isValid;
                     }
@@ -143,27 +140,7 @@ public class SubmitCollectionValidationService {
         if (fetchedItemEntity != null && incomingItemEntity.getBarcode().equals(fetchedItemEntity.getBarcode())) {
             isValid = true;
         } else {//Failure report - item id mismatch
-            SubmitCollectionReportInfo submitCollectionReportInfo = new SubmitCollectionReportInfo();
-            submitCollectionReportInfo.setItemBarcode(incomingItemEntity.getBarcode());
-            submitCollectionReportInfo.setCustomerCode(incomingItemEntity.getCustomerCode() != null ? incomingItemEntity.getCustomerCode() : "");
-            submitCollectionReportInfo.setOwningInstitution(owningInstitution);
-            ItemEntity misMatchedItemEntity = getMismatchedItemEntity(incomingItemEntity, fetchedOwningItemIdEntityMap);
-            if (misMatchedItemEntity != null) {
-                submitCollectionReportInfo.setMessage(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD + " - Owning institution item id mismatch - incoming owning institution item id " + incomingItemEntity.getOwningInstitutionItemId()
-                        + " , existing owning institution item id " + misMatchedItemEntity.getOwningInstitutionItemId()
-                        + ", existing owning institution holding id " + misMatchedItemEntity.getHoldingsEntities().get(0).getOwningInstitutionHoldingsId() + ", existing owning institution bib id "
-                        + misMatchedItemEntity.getBibliographicEntities().get(0).getOwningInstitutionBibId());
-                failureSubmitCollectionReportInfoList.add(submitCollectionReportInfo);
-            } else if(fetchedItemEntity==null){
-                List<ItemEntity> existingBarcodeDetails = itemDetailsRepository.findByBarcode(incomingItemEntity.getBarcode());
-                ItemEntity existingItemEntity = existingBarcodeDetails.get(0);
-                submitCollectionReportInfo.setOwningInstitution(existingItemEntity.getInstitutionEntity().getInstitutionCode());
-                submitCollectionReportInfo.setCustomerCode(existingItemEntity.getCustomerCode());
-                submitCollectionReportInfo.setItemBarcode(existingItemEntity.getBarcode());
-                submitCollectionReportInfo.setMessage(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD + " - Owning institution holdings id mismatch - incoming owning institution holdings id " + incomingOwningInstHoldingsId + ", existing owning institution item id " + existingItemEntity.getOwningInstitutionItemId()
-                        + ", existing owning institution holdings id " + existingItemEntity.getHoldingsEntities().get(0).getOwningInstitutionHoldingsId() + ", existing owning institution bib id : " + existingItemEntity.getBibliographicEntities().get(0).getOwningInstitutionBibId());
-                failureSubmitCollectionReportInfoList.add(submitCollectionReportInfo);
-            }
+            reportFailureRecords(incomingItemEntity, fetchedItemEntity, fetchedOwningItemIdEntityMap, incomingOwningInstHoldingsId, failureSubmitCollectionReportInfoList,owningInstitution);
             isValid = false;
         }
         return isValid;
@@ -216,15 +193,6 @@ public class SubmitCollectionValidationService {
             }
         }
         return null;
-    }
-
-    private String getExistingItemEntityOwningInstItemId(BibliographicEntity fetchedBibliographicEntity,ItemEntity incomingItemEntity){
-        for(ItemEntity fetchedItemEntity:fetchedBibliographicEntity.getItemEntities()){
-            if(fetchedItemEntity.getOwningInstitutionItemId().equals(incomingItemEntity.getOwningInstitutionItemId())){
-                return fetchedItemEntity.getHoldingsEntities().get(0).getOwningInstitutionHoldingsId();
-            }
-        }
-        return "";
     }
 
     public boolean validateIncomingItemHavingBibCountIsSameAsExistingItem(Map<String, List<SubmitCollectionReportInfo>> submitCollectionReportInfoMap, Map<String, ItemEntity> fetchedBarcodeItemEntityMap,
@@ -294,24 +262,25 @@ public class SubmitCollectionValidationService {
             isValidRecordToProcess &= false;
         } else if(holdingsIdUniqueList.size() > 1 &&
                 !Arrays.asList(nonHoldingIdInstitutionArray).contains(owningInstitutionCode)){//Owning inst Holdings id mismatch with in the incoming bound-with record
-            String multipleHoldingIds = holdingsIdUniqueList.stream().collect(Collectors.joining(","));
-            for(Map.Entry<String,String> owningBibIdOwnInstHoldingsIdEntry:owningBibIdOwnInstHoldingsIdMap.entrySet()){
-                StringBuilder message = new StringBuilder();
-                message.append(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD).append(RecapCommonConstants.HYPHEN).append("Incoming bound-with item has multiple owning institution holdings id attached to it, " +
-                        "multiple owning institution holdings are ").append(multipleHoldingIds).append(" - incoming owning institution holdings id ")
-                        .append(owningBibIdOwnInstHoldingsIdEntry.getValue()).append(", incoming owning institution item id ")
-                        .append(incomingBibliographicEntityList.get(0).getItemEntities().get(0).getOwningInstitutionItemId()).append(", ")
-                        .append("incoming owning institution bib id ").append(owningBibIdOwnInstHoldingsIdEntry.getKey());
-                String barcode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getBarcode();
-                String customerCode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getCustomerCode();
-                String owningInstitution = existingBibliographicEntityList.get(0).getItemEntities().get(0).getInstitutionEntity().getInstitutionCode();
-                submitCollectionReportHelperService.setSubmitCollectionReportInfo(submitCollectionReportInfoMap.get(RecapConstants.SUBMIT_COLLECTION_FAILURE_LIST),
-                        barcode,customerCode,owningInstitution,message.toString());
-            }
-            isValidRecordToProcess &= false;
+            isValidRecordToProcess = isValidRecordToProcess(submitCollectionReportInfoMap, incomingBibliographicEntityList, existingBibliographicEntityList, isValidRecordToProcess, holdingsIdUniqueList, owningBibIdOwnInstHoldingsIdMap);
         } /*else if(existingBibliographicEntityList.get(0).getOwningInstitutionBibId().substring(0, 1).equals("d")) {//validation to update existing dummy record if any (Removes existing dummy record and creates new record for the same barcode based on the input xml)
             isValidRecordToProcess &= true;
         }*/
+        return isValidRecordToProcess;
+    }
+
+    private boolean isValidRecordToProcess(Map<String, List<SubmitCollectionReportInfo>> submitCollectionReportInfoMap, List<BibliographicEntity> incomingBibliographicEntityList, List<BibliographicEntity> existingBibliographicEntityList, boolean isValidRecordToProcess, List<String> holdingsIdUniqueList, Map<String, String> owningBibIdOwnInstHoldingsIdMap) {
+        String multipleHoldingIds = holdingsIdUniqueList.stream().collect(Collectors.joining(","));
+        for (Map.Entry<String, String> owningBibIdOwnInstHoldingsIdEntry : owningBibIdOwnInstHoldingsIdMap.entrySet()) {
+            StringBuilder message = new StringBuilder();
+            message.append(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD).append(RecapCommonConstants.HYPHEN).append("Incoming bound-with item has multiple owning institution holdings id attached to it, " +
+                    "multiple owning institution holdings are ").append(multipleHoldingIds).append(" - incoming owning institution holdings id ")
+                    .append(owningBibIdOwnInstHoldingsIdEntry.getValue()).append(", incoming owning institution item id ")
+                    .append(incomingBibliographicEntityList.get(0).getItemEntities().get(0).getOwningInstitutionItemId()).append(", ")
+                    .append("incoming owning institution bib id ").append(owningBibIdOwnInstHoldingsIdEntry.getKey());
+            setSubmitCollectionReportInfo(submitCollectionReportInfoMap, existingBibliographicEntityList, message);
+        }
+        isValidRecordToProcess &= false;
         return isValidRecordToProcess;
     }
 
@@ -333,11 +302,7 @@ public class SubmitCollectionValidationService {
         if(!isItemAvailable) {
             StringBuilder message = new StringBuilder();
             message.append(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD).append(RecapCommonConstants.HYPHEN).append("Incoming record has reduced bib, but the bibs are not unlinked since the item is unavailable ");
-            String barcode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getBarcode();
-            String customerCode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getCustomerCode();
-            String owningInstitution = existingBibliographicEntityList.get(0).getItemEntities().get(0).getInstitutionEntity().getInstitutionCode();
-            submitCollectionReportHelperService.setSubmitCollectionReportInfo(submitCollectionReportInfoMap.get(RecapConstants.SUBMIT_COLLECTION_FAILURE_LIST),
-                    barcode,customerCode,owningInstitution,message.toString());
+            setSubmitCollectionReportInfo(submitCollectionReportInfoMap, existingBibliographicEntityList, message);
             isValidRecordToProcess &= false;
         }else if(matchedOwningInstBibIdList.size() > 0 && incomingBibsNotInExistingBibs.size() == 0 && holdingsIdUniqueList.size() == 1) {
             for (BibliographicEntity incomingBibliographicEntity : incomingBibliographicEntityList) {
@@ -351,31 +316,21 @@ public class SubmitCollectionValidationService {
             StringBuilder message = new StringBuilder();
             message.append(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD).append(RecapCommonConstants.HYPHEN).append("Incoming bound-with item with less bibs than the existing bibs which does not have matching bib that are available in the " +
                     "existing record, bib id(s) that are not linked with incoming item ").append(incomingBibsNotInExistingBibs.stream().collect(Collectors.joining(",")));
-            String barcode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getBarcode();
-            String customerCode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getCustomerCode();
-            String owningInstitution = existingBibliographicEntityList.get(0).getItemEntities().get(0).getInstitutionEntity().getInstitutionCode();
-            submitCollectionReportHelperService.setSubmitCollectionReportInfo(submitCollectionReportInfoMap.get(RecapConstants.SUBMIT_COLLECTION_FAILURE_LIST),
-                    barcode,customerCode,owningInstitution,message.toString());
+            setSubmitCollectionReportInfo(submitCollectionReportInfoMap, existingBibliographicEntityList, message);
             isValidRecordToProcess &= false;
         } else if(holdingsIdUniqueList.size() > 1 &&
                 !Arrays.asList(nonHoldingIdInstitutionArray).contains(owningInstitutionCode)){//Owning inst Holdings id mismatch with in the incoming bound-with record
-            String multipleHoldingIds = holdingsIdUniqueList.stream().collect(Collectors.joining(","));
-            for(Map.Entry<String,String> owningBibIdOwnInstHoldingsIdMapEntry:owningBibIdOwnInstHoldingsIdMap.entrySet()){
-                StringBuilder message = new StringBuilder();
-                message.append(RecapConstants.SUBMIT_COLLECTION_FAILED_RECORD).append(RecapCommonConstants.HYPHEN).append("Incoming bound-with item has multiple owning institution holdings id attached to it, " +
-                        "multiple owning institution holdings are ").append(multipleHoldingIds).append(" - incoming owning institution holdings id ")
-                        .append(owningBibIdOwnInstHoldingsIdMapEntry.getValue()).append(", incoming owning institution item id ")
-                        .append(incomingBibliographicEntityList.get(0).getItemEntities().get(0).getOwningInstitutionItemId()).append(", ")
-                        .append("incoming owning institution bib id ").append(owningBibIdOwnInstHoldingsIdMapEntry.getKey());
-                String barcode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getBarcode();
-                String customerCode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getCustomerCode();
-                String owningInstitution = existingBibliographicEntityList.get(0).getItemEntities().get(0).getInstitutionEntity().getInstitutionCode();
-                submitCollectionReportHelperService.setSubmitCollectionReportInfo(submitCollectionReportInfoMap.get(RecapConstants.SUBMIT_COLLECTION_FAILURE_LIST),
-                        barcode,customerCode,owningInstitution,message.toString());
-            }
-            isValidRecordToProcess &= false;
+            isValidRecordToProcess = isValidRecordToProcess(submitCollectionReportInfoMap, incomingBibliographicEntityList, existingBibliographicEntityList, isValidRecordToProcess, holdingsIdUniqueList, owningBibIdOwnInstHoldingsIdMap);
         }
         return isValidRecordToProcess;
+    }
+
+    private void setSubmitCollectionReportInfo(Map<String, List<SubmitCollectionReportInfo>> submitCollectionReportInfoMap, List<BibliographicEntity> existingBibliographicEntityList, StringBuilder message) {
+        String barcode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getBarcode();
+        String customerCode = existingBibliographicEntityList.get(0).getItemEntities().get(0).getCustomerCode();
+        String owningInstitution = existingBibliographicEntityList.get(0).getItemEntities().get(0).getInstitutionEntity().getInstitutionCode();
+        submitCollectionReportHelperService.setSubmitCollectionReportInfo(submitCollectionReportInfoMap.get(RecapConstants.SUBMIT_COLLECTION_FAILURE_LIST),
+                barcode, customerCode, owningInstitution, message.toString());
     }
 
     private List<String> getMatchedOwningInstBibId(List<BibliographicEntity> incomingBibliographicEntityList,List<BibliographicEntity> existingBibliographicEntityList){
@@ -455,6 +410,10 @@ public class SubmitCollectionValidationService {
 
     private String[] getNonHoldingIdInstitutionArray(){
         return nonHoldingIdInstitution.split(",");
+    }
+
+    private void reportFailureRecords(ItemEntity incomingItemEntity, ItemEntity fetchedItemEntity, Map<String, ItemEntity> fetchedOwningItemIdEntityMap, String incomingOwningInstHoldingsId, List<SubmitCollectionReportInfo> failureSubmitCollectionReportInfoList, String owningInstitution) {
+        submitCollectionReportHelperService.setFailureSubmitCollectionReportInfoList(failureSubmitCollectionReportInfoList, owningInstitution, fetchedOwningItemIdEntityMap, incomingOwningInstHoldingsId, incomingItemEntity, fetchedItemEntity);
     }
 
 }
